requirements.txt
requests
load_dotenv
flask
flask_sqlalchemy